organized functions of minirt_math and render_setup
