+++ src/render_setup.c
@@ -1,27 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   render.h                                           :+:      :+:    :+:   */
+/*   render_setup.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: davifah <dfarhi@student.42lausanne.ch      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/08/30 18:12:49 by davifah           #+#    #+#             */
-/*   Updated: 2022/09/01 11:22:00 by davifah          ###   ########.fr       */
+/*   Created: 2022/09/01 11:18:39 by davifah           #+#    #+#             */
+/*   Updated: 2022/09/01 11:18:53 by davifah          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef RENDER_H
-# define RENDER_H
+#include "minirt.h"
+#include "render.h"
 
-# include "minirt.h"
-# define LINE_BY_LINE_RENDER 1
+double	*get_angle_shift_per_pixel(unsigned char fov, int width, int height)
+{
+	double	*aspp;
 
-int			looper_mlx(void *param);
-t_vector	render_get_camera_direction(
-				t_vector *v, int x, int y);
-
-//render_setup
-double		*get_angle_shift_per_pixel(
-				unsigned char fov, int width, int height);
-
-#endif
+	aspp = ft_calloc(sizeof(double), 2);
+	if (width <= 1)
+		width = 2;
+	if (height <= 1)
+		height = 2;
+	aspp[0] = (double)fov / (width - 1);
+	aspp[1] = (double)fov / (height - 1);
+	return (aspp);
+}
